utils.py implemented for formating the code
